chore: added index file to configuration folder and improved filenames.
