+++ packages/shared/sdk-client/src/configuration/Configuration.test.ts
@@ -1,4 +1,4 @@
-import Configuration from './';
+import Configuration from './Configuration';
 
 describe('Configuration', () => {
   beforeEach(() => {
